pyproject.toml
[project]
name = "ddi-backend-template"
version = "0.1.1"
description = ""
readme = "README.md"
authors = [
    { name = "Dalberg Data Insights", email = "ddi.support@dalberg.com" },
]
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.0)",
    "uvicorn (>=0.32.0)",
    "sqlmodel (>=0.0.22,<0.0.23)",
    "psycopg2-binary (>=2.9.9)",
    "httpx (>=0.27.2)",
    "python-dotenv (>=1.0.1)",
    "pandas (>=2.2.3,<3.0.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.0"
pytest = "^8.0.0"
pre-commit = "^4.0.0"

[tool.ruff]
line-length = 99
target-version = "py311"

[tool.ruff.lint]
ignore = ["E203", "E501", "B008", "E266", "E501", "SIM", "I"]
select = ["B", "C", "E", "F", "W", "B9"]
fixable = ["ALL"]
unfixable = ["B"]

[tool.ruff.lint.pycodestyle]
max-doc-length = 99

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"